making it work the magic touch

+++ src/events/reactionAdd.ts
@@ -4,7 +4,7 @@ import {
   MessageReactionPayload,
   Reaction_Payload,
 } from "https://raw.githubusercontent.com/Skillz4Killz/Discordeno/master/types/message.ts";
-
+import { cache } from "https://raw.githubusercontent.com/Skillz4Killz/Discordeno/master/utils/cache.ts";
 function isMessage(
   message: Message | MessageReactionPayload,
 ): message is Message {
@@ -14,12 +14,18 @@ function isMessage(
 export const reactionAdd = async (
   message: Message | MessageReactionPayload,
   emoji: Reaction_Payload,
+  userID: string,
 ) => {
-  if (!isMessage(message)) return;
+  if (!isMessage(message) || !message.guild_id) return;
+
+  const guild = cache.guilds.get(message.guild_id);
+  if (!guild) return;
+
+  const reactingUser = cache.users.get(userID);
+  if (!reactingUser) return;
 
   if (emoji.name === "🔖") {
-    if (message.author.bot) return;
-
+    if (reactingUser.bot) return;
     // If the message does not have a issue/pull request number in it cancel.
     const number = message.content.split(" ").find((word) =>
       word.startsWith("#") && Number(word.substring(1))
@@ -29,15 +35,18 @@ export const reactionAdd = async (
     // A possible number was found
     const [issue, pull] = await Promise.all([
       fetch(
-        `https://api.github.com/repos/${configs.repositoryURL}/pulls/${number}`,
+        `https://api.github.com/repos/${configs.repositoryURL}/issues/${
+          number.substring(1)
+        }`,
       ).then((res) => res.json()),
       fetch(
-        `https://api.github.com/repos/${configs.repositoryURL}/pulls/${number}`,
+        `https://api.github.com/repos/${configs.repositoryURL}/pulls/${
+          number.substring(1)
+        }`,
       ).then((res) => res.json()),
     ]);
 
     let data;
-
     const isIssue = pull.message === "Not Found";
 
     if (issue.message !== "Not Found") data = issue;
@@ -62,7 +71,9 @@ export const reactionAdd = async (
           { name: "__**Status:**__", value: data.state, inline: true },
           {
             name: "__**Labels:**__",
-            value: data.labels.map((label: { name: any }) => label.name),
+            value: data.labels.map((label: { name: any }) => label.name).join(
+              ", ",
+            ) || "None",
             inline: true,
           },
         ],
@@ -83,6 +94,7 @@ export const reactionAdd = async (
     response.addReaction("🗑");
   } else if (emoji.name === "🗑") {
     if (message.author.id !== configs.botID) return;
+    if (userID === configs.botID) return;
     message.delete().catch(() => undefined);
   }
 };
